개행 및 줄맞춤 수정, Utils 저장위치 수정
개행 및 줄맞춤 수정하고 Utils 클래스 저장위치 수정했습니다.
