--- Assets/Prototype/Scripts/Utils.cs
@@ -1,34 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-
-[System.Serializable]
-public class PRS
-{
-    public Vector3 pos;
-    public Quaternion rot;
-    public Vector3 scale;
-
-    public PRS(Vector3 pos, Quaternion rot, Vector3 scale)
-    {
-        this.pos = pos;
-        this.rot = rot;
-        this.scale = scale;
-    }
-}
-
-public class Utils
-{
-    public static Quaternion QI => Quaternion.identity;
-
-    public static Vector3 MousePos
-    {
-        get
-        {
-            Vector3 result = Camera.main.ScreenToWorldPoint(Input.mousePosition);
-            result.z = -5;
-            return result;
-        }
-    }
-}
